+++ src/attentions/rfa.py
@@ -15,22 +15,22 @@ class RFA(Module):
     - Local sliding window attention for short-term context
     - Recurrent memory vector per head for long-term context
     """
-    def __init__(self, mp, n_ctx, n_emb, r_dropout, head_size, n_heads, window_size):
+    def __init__(self, mp, n_ctx, n_emb, r_dropout, s_head, n_heads, window_size):
         super().__init__()
-        assert head_size % n_heads == 0, "head_size must be divisible by n_heads"
+        assert s_head % n_heads == 0, "head_size must be divisible by n_heads"
         self.mp = mp
         self.n_ctx = n_ctx
         self.n_emb = n_emb
         self.r_dropout = r_dropout
-        self.head_size = head_size
+        self.s_head = s_head
         self.n_heads = n_heads
-        self.d_k = head_size // n_heads
+        self.d_k = s_head // n_heads
         self.window_size = window_size
 
-        self.q_proj = Linear(mp, n_emb, head_size, bias=False)
-        self.k_proj = Linear(mp, n_emb, head_size, bias=False)
-        self.v_proj = Linear(mp, n_emb, head_size, bias=False)
-        self.c_proj = Linear(mp, head_size, n_emb, bias=True)
+        self.q_proj = Linear(mp, n_emb, s_head, bias=False)
+        self.k_proj = Linear(mp, n_emb, s_head, bias=False)
+        self.v_proj = Linear(mp, n_emb, s_head, bias=False)
+        self.c_proj = Linear(mp, s_head, n_emb, bias=True)
 
         # Dropout layers
         self.attn_dropout = Dropout(mp, r_dropout)
@@ -76,7 +76,7 @@ class RFA(Module):
         flops = 0
 
         # Q, K, V projections
-        flops += 3 * batch_size * self.n_ctx * self.n_emb * self.head_size * 2
+        flops += 3 * batch_size * self.n_ctx * self.n_emb * self.s_head * 2
 
         # Attention score computation: Q @ K^T (local window)
         flops += batch_size * self.n_heads * self.n_ctx * self.window_size * self.d_k * 2
@@ -91,7 +91,7 @@ class RFA(Module):
         flops += batch_size * self.n_heads * self.n_ctx * self.window_size * self.d_k * 2
 
         # Output projection
-        flops += batch_size * self.n_ctx * self.head_size * self.n_emb * 2
+        flops += batch_size * self.n_ctx * self.s_head * self.n_emb * 2
 
         # Bias add for output projection
         if self.c_proj.bias is not None: